# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: NBA API
  version: 1.0.0
  description: Context-Aware Next Best Action API
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /doctors:
    get:
      summary: List doctors
      parameters:
        - in: query
          name: tier
          schema: { type: string }
        - in: query
          name: specialty
          schema: { type: string }
        - in: query
          name: territoryId
          schema: { type: string, format: uuid }
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDoctor'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create doctor
      responses:
        '201':
          description: Created
  /doctors/{id}:
    get:
      summary: Get doctor
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Doctor'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /doctors/nearby:
    get:
      summary: Find nearby doctors
      parameters:
        - in: query
          name: lat
          required: true
          schema: { type: number }
        - in: query
          name: lon
          required: true
          schema: { type: number }
        - in: query
          name: radiusKm
          schema: { type: number, default: 5 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Doctor'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        path: { type: string }
        timestamp: { type: string, format: date-time }
...
    Doctor:
      type: object
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
        specialty: { type: string }
        tier: { type: string }
        priorityScore: { type: integer }
        territoryId: { type: string, format: uuid }
        notes: { type: string }
    PageDoctor:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Doctor'
        meta:
          $ref: '#/components/schemas/PageMeta'
    PageMeta:
      type: object
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
...